refactor: 패키지명 model -> domain 수정
