+++ src/main/java/racingcar/domain/CarCondition.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.Random;
