+++ src/main/java/racingcar/domain/nickname/CanNotNullNicknameException.java
@@ -1,4 +1,4 @@
-package racingcar.model.nickname;
+package racingcar.domain.nickname;
 
 public class CanNotNullNicknameException extends RuntimeException {
