+++ src/test/java/racingcar/domain/CarsTest.java
@@ -1,9 +1,9 @@
-package racingcar.model;
+package racingcar.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.model.nickname.Nickname;
+import racingcar.domain.nickname.Nickname;
 
 import java.util.List;
 import java.util.stream.Collectors;
