+++ src/test/java/racingcar/domain/FailCondition.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class FailCondition implements Condition {
