+++ src/test/java/racingcar/domain/SuccessCondition.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class SuccessCondition implements Condition {
